Update README + LICENSE file.md
🎬 YouTube Clone – My HTML & CSS Journey
This project is my take on recreating the look and feel of YouTube only featuring with HTML and CSS.
It’s not just a clone – it’s my personal playground where I experimented with design, responsiveness, and some fun visual effects to make it feel alive. This project demonstrates responsive design using Flexbox and modern CSS styling.